Loader splitted for web
